% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanImputationParam-class.R
\docType{class}
\name{MeanImputationParam-class}
\alias{MeanImputationParam-class}
\title{S4 class for the parameters of the mean imputation method}
\description{
Definition of the S4 class named \code{MeanImputationParam} for the parameters for
the mean imputation method of missing values.
}
\section{Slots}{

\describe{
\item{\code{VarNames}}{character vector with the names of the variables to impute.}

\item{\code{DomainNames}}{character vector with the names of the variables determining the domains over
which to compute the mean.}
}}

\examples{
# An empty MeanImputationParam object:
new(Class = 'MeanImputationParam')

}
