% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedianImputationParam-class.R
\docType{class}
\name{MedianImputationParam-class}
\alias{MedianImputationParam-class}
\title{S4 class for the parameters of the median imputation method}
\description{
Definition of the S4 class named \code{MedianImputationParam} for the parameters for
the median imputation method of missing values.
}
\section{Slots}{

\describe{
\item{\code{VarNames}}{character vector with the names of the variables to impute.}

\item{\code{DomainNames}}{character vector with the names of the variables determining the domains over
which to compute the median.}
}}

\examples{
# An empty MedianImputationParam object:
new(Class = 'MedianImputationParam')

}
