% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setVarNames.R
\name{setVarNames<-}
\alias{setVarNames<-}
\title{Set value of component \code{VarNames} of an object \linkS4class{ImputationParam}}
\usage{
setVarNames(object) <- value
}
\arguments{
\item{object}{Object \linkS4class{ImputationParam} whose component \code{VarNames} is to be
assigned.}

\item{value}{character vector to be assigned to the component \code{VarNames}.}
}
\value{
Object \linkS4class{ImputationParam} with the component \code{VarNames} updated.
}
\description{
\code{setVarNames} assigns a character value to the component \code{VarNames}  of
the input object \linkS4class{ImputationParam}.
}
