% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\docType{methods}
\name{Impute}
\alias{Impute}
\alias{Impute,data.table,MeanImputationParam-method}
\title{Apply the imputation method to a data set}
\usage{
Impute(object, Param)

\S4method{Impute}{data.table,MeanImputationParam}(object, Param)
}
\arguments{
\item{object}{Object containing the the data upon which the imputation is to be computed.}

\item{Param}{Object of virtual class \linkS4class{ImputationParam} with the parameters specifying
the method of imputation to be applied.}
}
\value{
Object of the same class as the input \code{object} with the missing values imputed
according to the method specified in the input argument \code{Param}
}
\description{
Apply the imputation method to a data set
}
\examples{
\dontrun{}

}

